Add support to set spower superior account (#943)
Add support to set spower superior account.

This is a pre merge commit for the offchain spower calculation feature. We will upgrade this runtime first to set the spower superior beforehand, so that after we upgrade the actual runtime which contains the offchain spower feature, the crust-spower service can run directly without any wait time.